Add private address::error module
In preparation for improving error handling in the `address` module move
`address.rs` to `address/mod.rs`, add an `address/error.rs` file, and
move the public error type to the new module.

Note:

- Make there module private, this is a code organization thing only.
- Use wildcard re-export of `self::error::*` in `address/mod.rs`
- Use `super` to import things from `address` into `error`

Refactor only, no logic changes.
